Add contributors to credits
As Termora doesn't have the server members intent, and as such, cannot
fetch the member list of *any* guild, this also adds a helper bot that
fetches the member list for the main process.  The helper communicates
with the main process via RPC.  Right now, it only handles server
members, but once the message content intent rolls out, it will likely
handle that too.